Make Attribute::setName() protected
It must not be allowed to change the name of an attribute after it is
created. Attribute instances are usually created and registered via the
exposed methods of the Attributes class which does not take changes of
Attribute::setName() into account, e.g.

$attributes = (new Attributes())
    ->add('name', 'test');

$attributes->get('name')->setName('enabled');

$attributes->get('enabled');

The last call would implicitly create a new Attribute because the name
Attribute is still registered with name. Now we have two attributes
with the name enabled with ambiguous state.