Introduce FormElementContainer::$defaultElementDecoratorLoaderRegistered
The method ensureDefaultElementDecoratorLoaderRegistered (before
eventually...) is here to register the default element decorator loader
if it has not been loaded before. Despite having the possibility to
register custom loaders, we always want to register our loader as well.
Before, this method just checked if any loader was registered. If a
subclass calls addPluginLoader before this method is called, our default
decorator loader would not be registered. Now we use a variable which
protects us from that.

Please note that our default decorator should always be the last of all
loaders so that it is possible for custom loaders to provide alternative
implementations for our default decorators. But this is not subject to
this commit.