Introduce FormElementContainer::$defaultElementLoaderRegistered
The method ensureDefaultElementLoaderRegistered (before eventually...)
is here to register the default element loader if it has not been
registered before. Despite having the possibility to register custom
loaders, we always want to register our loader as well. Before, this
method just checked if any loader was registered. If
addPluginLoader() is somehow called before this method, our default
element loader would not get registered. Now we use a variable
which protects us from that.

Please note that our default loader should always be the last of all
loaders so that it is possible for custom loaders to provide alternative
implementations for our default elements. But this is not subject to
this commit.