FormElementContainer: Refine add/create/registerElement() flow
addElement() is now the only method which supports providing either type
plus name or an element instance. Thus registerElement now expects an
element instance.
So, addElement() is the common entrypoint and createElement() plus
registerElement() can still be used in order to control the element
rendering. This is much cleaner code and API wise.

addElement() and registerElement() now have security checks in place so
that a name is always required.

Before, it was possible to register the very same element with different
names. This is no longer possible.