CheckboxElement: Render hidden element for the unchecked value
A hidden element is required because unchecked checkboxes are
not submitted. Without a submitted value it's difficult to
invalidate a default value. The hidden element is always
submitted but overriden by the checkbox in case it's checked.
This has also the side-effect that `getValue()` now provides
a meaningful return value. (i.e. not always the checked value)