Also set object ID (to rebind callbacks) on assemble
Currently the object ID is only set if the element has attributes. This
made sense since callbacks only need to be rebound after cloning the
element if attributes are set. However, with upcoming changes we will
also rebind callbacks of all content elements so that the object id
needs to be available in cases where the element has no attributes but
content.