Form: Trigger `ON_REQUEST` only in case it's not been sent
Literally no one ever used this event, since it served
absolutely no purpose. Only a single control registers
it on its own: ipl-web's `SortControl`, but guess what?
The first check is whether the form has been sent or not
and the listener only does something if not.

And of course, it still needs to be triggered before
assembly, as a listener will probably influence it.