Deprecate interface Wrappable
The implementation of this interface in the HtmlDocument class requires the property of the class (HtmlDocument::renderWrapped() uses the renderedBy property).
This means that a Wrappable instance must actually also be an HtmlDocument instance. This in turn means that Wrappable can be removed (depricated) and the HtmlDocument can be used directly instead.