FormDecorator: Do not expect form elements to be HTML elements
This was a mis-conception made during #146. `FormElement` may
provide `getAttributes()`, but this was only added because of
options. Without this, the remaining interface does not describe
a HTML element at all. There is even a case, I stumbled over today,
where this is alread a thing: `ipl\Html\FormElement\RadioElement`
This does not have a tag, and overrides `renderUnwrapped` in order
to not having to define one. Though, some of the new decoators
also call `getTag()` and so they fail while decorating such an
element. This will be tackled in a follow-up PR though.