DecoratorChain: Allow to replace specific decorators
For this all decorators require an identifier now.
This is a breaking change compared to the current
state. In the past using Zend form, we often copied
the default list of decorators just to change a single
decorator. Though, changes to the default decorators
were never applied by forms overriding them, so this
functionality in the new implementation should provide
more flexibility in this regard.